+++ electron/controllers/ChatController.ts
@@ -1,0 +1,21 @@
+export class ChatController {
+  static async getChats() {
+    return 'getChats';
+  }
+
+  static async getChat() {
+    return 'getChat';
+  }
+
+  static async updateChat() {
+    return 'updateChat';
+  }
+
+  static async createNewChat() {
+    return 'createNewChat';
+  }
+
+  static async deleteChat() {
+    return 'deleteChat';
+  }
+};