+++ electron/controllers/ModelController.ts
@@ -1,0 +1,15 @@
+// import { IpcMainInvokeEvent } from "electron";
+
+export class ModelController {
+  static async getModels() {
+    return 'getModels';
+  }
+
+  static async downloadModel() {
+    return 'downloadModel';
+  }
+
+  static async deleteModel() {
+    return 'deleteModel';
+  }
+};