+++ electron/providers/routerProvider.ts
@@ -1,0 +1,7 @@
+import { startRouter } from "./../routes";
+
+export class RouterProvider {
+  static init() {
+    startRouter();
+  }
+}