+++ src/mock/modelFile.ts
@@ -4,33 +4,49 @@ export const mockModelFiles: TModelFile[] = [
   {
     id: 'llama 1.0',
     name: 'Llama 1.0',
+    url: 'https://llama.com/llama-1.0',
+    downloadUrl: 'https://llama.com/llama-1.0',
   },
   {
     id: 'llama 2.0',
     name: 'Llama 2.0',
+    url: 'https://llama.com/llama-1.0',
+    downloadUrl: 'https://llama.com/llama-1.0',
   },
   {
     id: 'llama 2.1',
     name: 'Llama 2.1',
+    url: 'https://llama.com/llama-1.0',
+    downloadUrl: 'https://llama.com/llama-1.0',
   },
   {
     id: 'llama 3.0',
     name: 'Llama 3.0',
+    url: 'https://llama.com/llama-1.0',
+    downloadUrl: 'https://llama.com/llama-1.0',
   },
   {
     id: 'llama 3.1',
     name: 'Llama 3.1',
+    url: 'https://llama.com/llama-1.0',
+    downloadUrl: 'https://llama.com/llama-1.0',
   },
   {
     id: 'llama 3.2',
     name: 'Llama 3.2',
+    url: 'https://llama.com/llama-1.0',
+    downloadUrl: 'https://llama.com/llama-1.0',
   },
   {
     id: 'llama 3.3',
     name: 'Llama 3.3',
+    url: 'https://llama.com/llama-1.0',
+    downloadUrl: 'https://llama.com/llama-1.0',
   },
   {
     id: 'llama 3.4',
     name: 'Llama 3.4',
+    url: 'https://llama.com/llama-1.0',
+    downloadUrl: 'https://llama.com/llama-1.0',
   },
 ];