+++ src/plugins/ipc/initIpcLog.ts
@@ -7,6 +7,8 @@ export const initIpcLog = () => {
     return;
   }
 
+  console.log('initIpcLog');
+
   window.ipcRenderer.on('main-process-message', (_event, message) => {
     console.log(message);
   });